refactor: add custom base class
